Try to make the "Save the key" button work more reliably
Remove the stupid "set form target to invisible iframe" kludge *and*
the "id" attribute from the username input. They were both preventing the
"save ..." button from working in Chrome. When submitting a form,
Chrome checks if the new page contains a form similar to the one it was
sent there from, and if it finds such a form, it considers that the
login was "unsuccesful" and the password doesn't need saving. A form
without an id attribute on its "user" input doesn't qualify as similar
enough ;-)

Also add the autocomplete=username which is apparently needed on safari.

Since the page is now actually *navigating away* (to the current location
with a "?" appended), add a notice to help in the case where the password
manager still didn't pop up the "save password" dialog and the entered
user + key appear "lost".