merge page.html and popup.html into a single file; use css to hide buttons, labels and extra text when shown in the popup window
Append a '#popup' hash in manifest.json to the popup's path and use
that to adjust the css so that only the "user" form input, the "code"
div and the "add key" link are shown.

Use a simple clickable div instead of a checkbox to switch the key
input from password to visible text.

Don't add the PasswordCredential chunk to the xpi -- despite being
conditional, it may trigger spurious compat warnings; and it's quite
useless, anyways, since firefox doesn't implement that API yet.
Not even in nightly.