use the already existing #id named variables instead of querySelector(#id)
for each element with "id=foo", the browser is creating a window property
(= global variable) with the same name referring to that element.

that's a standard feature, not an obscure hack:
https://html.spec.whatwg.org/multipage/nav-history-apis.html#named-access-on-the-window-object

use that feature, making the id's uppercase so they stand out and do not
clash with existing function names or variables.