selectAllChildren() no longer works in X11 firefox >= 121.0a1
Since the nightly half-fix[1] for the pastejack bug[2] I had reported,
selectAllChildren() no longer sets the primary selection on X11
in firefox.

So just prevent the deceptive highlighting on gecko >= 121.

The "copied!" text would still LIE, since the text will only
be copied to the windows-style clipboard. But there's no point
in also breaking it for the Linux users who are not using the
primary selection, just for the sake of keeping things consistent.

People would still be able to copy the generated code to the primary
selection with a single click on the code or the "click to copy"
message, because of the "user-select: all" css rule.

[1] https://hg.mozilla.org/mozilla-central/rev/88e0043c5aa4234dada941ac2fd0ded875210508
[2] https://github.com/turistu/odds-n-ends/blob/main/firefox/pastejack.md